requirements.txt
discord.py
requests
psutil
time
asyncio
random